osa4: 4.15-4.23

+++ tests/blogs_api.test.js
@@ -1,41 +1,11 @@
 const mongoose = require('mongoose')
 const supertest = require('supertest')
 const Blog = require('../models/blog')
+const {initialBlogs, testBlog,testBlogWithoutLikes,testBlogWithoutUrlAndTitle,initTestBlogs} = require('./testData')
 const app = require('../app')
 
 const api = supertest(app)
 
-const initialBlogs = [
-    {
-      title: 'Go To Statement Considered Harmful',
-      author: 'Edsger W. Dijkstra',
-      url: 'http://www.u.arizona.edu/~rubinson/copyright_violations/Go_To_Considered_Harmful.html',
-      likes: 5,
-    },
-    {
-        title: "React patterns",
-        author: "Michael Chan",
-        url: "https://reactpatterns.com/",
-        likes: 7,
-    },
-]
-const testBlog =     {
-    title: "Testing blogs",
-    author: "Blog Tester",
-    url: "https://blogtester.com/",
-    likes: 1,
-}
-
-const testBlogWithoutLikes =     {
-    title: "Testing blogs",
-    author: "Blog Tester",
-    url: "https://blogtester.com/"
-}
-
-const testBlogWithoutUrlAndTitle =     {
-    author: "Blog Tester",
-    likes: 1
-}
 describe('blogs tests',() =>{
 
 
@@ -50,8 +20,11 @@ describe('blogs tests',() =>{
         expect(response.body[1].id).toBeDefined()
         expect(response.statusCode).toBe(200)  
     })
-    test('when user makes a POST request to /api/blogs with valid input new blog should be added', async() =>{
-        const responsePOST = await api.post('/api/blogs').send(testBlog)
+    test('when user makes a POST request to /api/blogs with valid input a new blog should be added', async() =>{
+        const loginResponse = await api.post('/api/login').send({username:'Pepe',password:'Pepe'})
+        const token = loginResponse.body.token;
+
+        const responsePOST = await api.post('/api/blogs').send(testBlog).set({Authorization: `Bearer ${token}`})
         expect(responsePOST.statusCode).toBe(201) 
         const responseGET = await api.get('/api/blogs')
 
@@ -64,7 +37,9 @@ describe('blogs tests',() =>{
     })
 
     test('when user makes a POST request to /api/blogs with input without likes a new blog should be added with zero likes', async() =>{
-        const responsePOST = await api.post('/api/blogs').send(testBlogWithoutLikes)
+        const loginResponse = await api.post('/api/login').send({username:'Pepe',password:'Pepe'})
+        const token = loginResponse.body.token;
+        const responsePOST = await api.post('/api/blogs').send(testBlogWithoutLikes).set({Authorization: `Bearer ${token}`})
         expect(responsePOST.statusCode).toBe(201) 
         const responseGET = await api.get('/api/blogs')
         expect(responseGET.body).toHaveLength(3)
@@ -75,29 +50,45 @@ describe('blogs tests',() =>{
         expect(responseGET.statusCode).toBe(200)  
     })
     test('when user makes a POST request to /api/blogs with input without title and url a new blog should not be added, instead 400 statuscode is returned', async() =>{
-        const responsePOST = await api.post('/api/blogs').send(testBlogWithoutUrlAndTitle)
+        const loginResponse = await api.post('/api/login').send({username:'Pepe',password:'Pepe'})
+        const token = loginResponse.body.token;
+        const responsePOST = await api.post('/api/blogs').send(testBlogWithoutUrlAndTitle).set({Authorization: `Bearer ${token}`})
         expect(responsePOST.statusCode).toBe(400) 
         const responseGET = await api.get('/api/blogs')
         expect(responseGET.body).toHaveLength(initialBlogs.length)
         expect(responseGET.statusCode).toBe(200)  
     })
+    test('when user makes a POST request to /api/blogs with valid input, but without  a token no blog should be created and 401 is returned', async() =>{
+        const responsePOST = await api.post('/api/blogs').send(testBlog)
+        expect(responsePOST.statusCode).toBe(401) 
+        const responseGET = await api.get('/api/blogs')
+        expect(responseGET.body).toHaveLength(2)  
+    })
     test('when user makes a DELETE request to /api/blogs/:id with existing blog\'s id, the blog is deleted and 204 is returned', async() =>{
+        /**FIX THIS */
+        const loginResponse = await api.post('/api/login').send({username:'Pepe',password:'Pepe'})
+        const token = loginResponse.body.token;
         const responseGET = await api.get('/api/blogs')
         expect(responseGET.statusCode).toBe(200) 
-        const responseDELETE = await api.delete(`/api/blogs/${responseGET.body[0].id}`)
+        const responseDELETE = await api.delete(`/api/blogs/${responseGET.body[0].id}`).send(testBlogWithoutLikes).set({Authorization: `Bearer ${token}`})
         expect(responseDELETE.statusCode).toBe(204) 
         const responseGET2 = await api.get('/api/blogs')
         expect(responseGET2.body).toHaveLength(1)
         expect(responseGET2.statusCode).toBe(200) 
     })
     test('when user makes a DELETE request to /api/blogs/:id with non-existing blog\'s id, 404 is returned and blogs state stays the same', async() =>{
-        const responseDELETE = await api.delete('/api/blogs/000000000000000000000000')
+        const loginResponse = await api.post('/api/login').send({username:'Pepe',password:'Pepe'})
+        const token = loginResponse.body.token;
+        const responseDELETE = await api.delete('/api/blogs/000000000000000000000000').set({Authorization: `Bearer ${token}`})
         expect(responseDELETE.statusCode).toBe(404) 
         const responseGET = await api.get('/api/blogs')
         expect(responseGET.body).toHaveLength(2)
         expect(responseGET.statusCode).toBe(200) 
     })
     test('when user makes a PUT request to /api/blogs/:id with existing blog\'s id and valid input the blog entry is updated', async() =>{
+        /**FIX THIS */
+        const loginResponse = await api.post('/api/login').send({username:'Pepe',password:'Pepe'})
+        const token = loginResponse.body.token;
         const responseGET = await api.get('/api/blogs')
         expect(responseGET.statusCode).toBe(200)
         const updatedBlog = {
@@ -106,7 +97,7 @@ describe('blogs tests',() =>{
             url: responseGET.body[0].url,
             likes: responseGET.body[0].likes+1
         }
-        const responsePUT = await api.put(`/api/blogs/${responseGET.body[0].id}`).send(updatedBlog)
+        const responsePUT = await api.put(`/api/blogs/${responseGET.body[0].id}`).send(updatedBlog).set({Authorization: `Bearer ${token}`})
         expect(responsePUT.statusCode).toBe(204) 
         const responseGET2 = await api.get('/api/blogs')
         expect(responseGET2.statusCode).toBe(200) 
@@ -114,17 +105,13 @@ describe('blogs tests',() =>{
         
     })
     test('when user makes a PUT request to /api/blogs/:id with non-existing blog\'s id, 404 is returned', async() =>{
-        const responsePUT = await api.put('/api/blogs/000000000000000000000000').send(testBlog)
+        const loginResponse = await api.post('/api/login').send({username:'Pepe',password:'Pepe'})
+        const token = loginResponse.body.token;
+        const responsePUT = await api.put('/api/blogs/000000000000000000000000').send(testBlog).set({Authorization: `Bearer ${token}`})
         expect(responsePUT.statusCode).toBe(404) 
     })
 })
-beforeEach(async () => {
-    await Blog.deleteMany({})
-    let blogObject = new Blog(initialBlogs[0])
-    await blogObject.save()
-    blogObject = new Blog(initialBlogs[1])
-    await blogObject.save()
-  })
+beforeEach(initTestBlogs)
 afterAll(()=>{
     mongoose.connection.close()
 })
